notebooks/handwriting-recognition-mnist/mnist-webui/requirements.txt
flask 
numpy 
Pillow 
requests